add x-axis compensation and delete unneeded features
