+++ package.xml
@@ -2,7 +2,7 @@
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
   <name>object_detection_avix</name>
-  <version>0.0.0</version>
+  <version>1.0.1</version>
   <description>TODO: Package description</description>
   <maintainer email="alexmanson_lu@outlook.com">avix</maintainer>
   <license>TODO: License declaration</license>
