+++ setup.py
@@ -25,8 +25,7 @@ setup(
     tests_require=['pytest'],
     entry_points={
         'console_scripts': [
-            'tracking_node = object_detection_avix.tracking_node_v6:main',
-            'test_time=object_detection_avix.test_time:main'
+            'tracking_node = object_detection_avix.tracking_node:main',
         ],
     },
 )
